// src/entities/MentalStats/model/contributionText.ts
// Временное решение. По факту каждое воздействие должно иметь свои уникальные ответы, которые
//зависят от того, что мы делаем (например, наш подопечный съедает шоколадку), от состояния 
//подопечного (например, подопечный радуется, грустит или вовсе не любит шоколад)
// и от истории наград/наказаний (Что? Всю неделю дают шоколад?)
export interface ContributionText {
    InsufficientReward: string, //Недостаточное вознаграждение
    ExcessiveReward: string, //Достаточное вознаграждение
    SufficientReward: string, //Избыточное вознаграждение
    UndeservedReward: string,  //Недостаточное вознаграждение
    UnfairReward: string,  //Несправедливое - это когда требовалось наказание, а получили вознаграждение.
    InsufficientPunishment: string,// Недостаточное наказание
    SufficientPunishment: string, // Достаточное наказание
    ExcessivePunishment: string, // Избыточное наказание
    UndeservedPunishment: string, // Незаслуженное наказание
    UnfairPunishment: string, // Несправедливое наказание
}

export const contributionText: ContributionText = {
    InsufficientReward: "Работник улыбается, чтобы показать свою признательность за оказанную милость, но вы замечаете, что его глаза задумчиво и беспомощно смотрят в никуда",
    ExcessiveReward: "В результате работник приободрился.",
    SufficientReward: "В результате работник заметно приободрился и даже начал улыбаться. Ух ты. Улыбка. У нас здесь.",
    UndeservedReward: "В результате работник радостно улыбается, хотя иногда кажется, что он имеет какой-то невысказанный вопрос.",
    UnfairReward: "Работник воспринимает оказанную милость с нексрываемым удивлением. Кажется, ваши воспитательные методы вызывают вопросы не только у вас самих.",
    InsufficientPunishment: "Работник понуро опускает голову и извиняется, но, кажется, он недостаточно впечатлён вашей силой.",
    SufficientPunishment: "Работник понуро опускает голову и обещает, что такое больше не повторится.",
    ExcessivePunishment: "Работник плачет",
    UndeservedPunishment: "Вы видите нотки обиды у работника. Он даже ничего ещё не сделал, а уже получил.",
    UnfairPunishment: "Работник очень не согласен с таким отношением к себе. Кажется, он серьёзно подумывает о том, чтобы закончить свою жизнь как служащего вашей гнилой компании."
};